Removed dangerous condition in TV help compiler.
The code has some ambiguous condition, reported by gcc.
I first added parenthesis to make it behave just like it does
without the parenthesis. But the GitHub code I'm analyzing used
a different approach. But I'm not sure if the code was
originally wrongly coded or had another intention.
As the condition involves the skip of ASCII 8 (backspace) I'm
just removing this posibility.