Added comment about MSVC potential need of internal usleep.
I saw it as part of the GitHub patch:
windoze/tvision@6ac86c30f2e23a84df0bedd9b6c98ba49a548d26

But I don't know if this is really needed. Is added at a level
that affects all the MSVC versions and the code compiled for
all MSVCs from 4 to 8, so I can't see why could be needed for
all.