Added some rudimentary CMake support.
From GitHub project windoze/tvision
I'm not sure which targets are benefited from this.
For Linux this is just a toy because it only generate a static
library, which can be used to generate big binaries that doesn't
suport the X11 driver.
May be can help Windows targets, so I'm keeping the files.
In the future they could be the base for a real CMake support.